wasm_bindgen

Module closure

Support for long-lived closures in wasm-bindgen

This module defines the Closure type which is used to pass “owned closures” from Rust to JS. Some more details can be found on the Closure type itself.

Structs

Closure

A handle to both a closure in Rust as well as JS closure which will invoke the Rust closure.