Trait wasm_bindgen::convert::IntoWasmAbi

pub trait IntoWasmAbi: WasmDescribe {
    type Abi: WasmAbi;
    fn into_abi(self, extra: &mut dyn Stack) -> Self::Abi;
}

A trait for anything that can be converted into a type that can cross the wasm ABI directly, eg u32 or f64.

This is the opposite operation as FromWasmAbi and Ref[Mut]FromWasmAbi.

Associated Types

type Abi: WasmAbi

The wasm ABI type that this converts into when crossing the ABI boundary.

Required Methods

fn into_abi(self, extra: &mut dyn Stack) -> Self::Abi

Convert self into Self::Abi so that it can be sent across the wasm ABI boundary.