wasm_bindgen::convert

Trait ReturnWasmAbi

pub trait ReturnWasmAbi: WasmDescribe {
    type Abi: WasmAbi;

    // Required method
    fn return_abi(self) -> Self::Abi;
}

A trait representing how to interpret the return value of a function for the Wasm ABI.

This is very similar to the IntoWasmAbi trait and in fact has a blanket implementation for all implementors of the IntoWasmAbi. The primary use case of this trait is to enable functions to return Result, interpreting an error as “rethrow this to JS”

⚠️ Unstable

This is part of the internal convert module, no stability guarantees are provided. Use at your own risk. See its documentation for more details.

Required Associated Types

type Abi: WasmAbi

Same as IntoWasmAbi::Abi

Required Methods

fn return_abi(self) -> Self::Abi

Same as IntoWasmAbi::into_abi, except that it may throw and never return in the case of Err.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.