Struct wasm_bindgen::js::Array

pub struct Array { /* fields omitted */ }

Methods

impl Array

pub fn copy_within(&self, target: i32, start: i32, end: i32) -> Array

The copyWithin() method shallow copies part of an array to another location in the same array and returns it, without modifying its size.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/copyWithin

impl Array

pub fn concat(&self, array: &Array) -> Array

The concat() method is used to merge two or more arrays. This method does not change the existing arrays, but instead returns a new array.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/concat

impl Array

pub fn every(&self, predicate: &mut FnMut(JsValue, u32, Array) -> bool) -> bool

The every() method tests whether all elements in the array pass the test implemented by the provided function.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/every

impl Array

pub fn fill(&self, value: JsValue, start: u32, end: u32) -> Array

The fill() method fills all the elements of an array from a start index to an end index with a static value. The end index is not included.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/fill

impl Array

pub fn filter(
&self,
predicate: &mut FnMut(JsValue, u32, Array) -> bool
) -> Array

The filter() method creates a new array with all elements that pass the test implemented by the provided function.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter

impl Array

pub fn find(
&self,
predicate: &mut FnMut(JsValue, u32, Array) -> bool
) -> JsValue

The find() method returns the value of the first element in the array that satisfies the provided testing function. Otherwise undefined is returned.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/find

impl Array

pub fn includes(&self, value: JsValue, from_index: i32) -> bool

The includes() method determines whether an array includes a certain element, returning true or false as appropriate.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/includes

impl Array

pub fn index_of(&self, value: JsValue, from_index: i32) -> i32

The indexOf() method returns the first index at which a given element can be found in the array, or -1 if it is not present.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/indexOf

impl Array

pub fn is_array(value: &JsValue) -> bool

The Array.isArray() method determines whether the passed value is an Array.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/isArray

impl Array

pub fn join(&self, delimiter: &str) -> JsString

The join() method joins all elements of an array (or an array-like object) into a string and returns this string.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/join

impl Array

pub fn last_index_of(&self, value: JsValue, from_index: i32) -> i32

The lastIndexOf() method returns the last index at which a given element can be found in the array, or -1 if it is not present. The array is searched backwards, starting at fromIndex.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/lastIndexOf

impl Array

pub fn length(&self) -> u32

The length property of an object which is an instance of type Array sets or returns the number of elements in that array. The value is an unsigned, 32-bit integer that is always numerically greater than the highest index in the array.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/length

impl Array

pub fn map(
&self,
predicate: &mut FnMut(JsValue, u32, Array) -> JsValue
) -> Array

map calls a provided callback function once for each element in an array, in order, and constructs a new array from the results. callback is invoked only for indexes of the array which have assigned values, including undefined. It is not called for missing elements of the array (that is, indexes that have never been set, which have been deleted or which have never been assigned a value).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/map

impl Array

pub fn pop(&self) -> JsValue

The pop() method removes the last element from an array and returns that element. This method changes the length of the array.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/pop

impl Array

pub fn push(&self, value: JsValue) -> u32

The push() method adds one or more elements to the end of an array and returns the new length of the array.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/push

impl Array

pub fn reverse(&self) -> Array

The reverse() method reverses an array in place. The first array element becomes the last, and the last array element becomes the first.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/reverse

impl Array

pub fn shift(&self) -> JsValue

The shift() method removes the first element from an array and returns that removed element. This method changes the length of the array.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/shift

impl Array

pub fn slice(&self, start: u32, end: u32) -> Array

The slice() method returns a shallow copy of a portion of an array into a new array object selected from begin to end (end not included). The original array will not be modified.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice

impl Array

pub fn some(&self, predicate: &mut FnMut(JsValue) -> bool) -> bool

The some() method tests whether at least one element in the array passes the test implemented by the provided function. Note: This method returns false for any condition put on an empty array. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/some

impl Array

pub fn sort(&self) -> Array

The sort() method sorts the elements of an array in place and returns the array. The sort is not necessarily stable. The default sort order is according to string Unicode code points.

The time and space complexity of the sort cannot be guaranteed as it is implementation dependent.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort

impl Array

pub fn to_string(&self) -> JsString

The toString() method returns a string representing the specified array and its elements.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/toString

impl Array

pub fn unshift(&self, value: JsValue) -> u32

The unshift() method adds one or more elements to the beginning of an array and returns the new length of the array.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/unshift

impl Array

pub fn keys(&self) -> ArrayIterator

The keys() method returns a new Array Iterator object that contains the keys for each index in the array.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/keys

impl Array

pub fn entries(&self) -> ArrayIterator

The entries() method returns a new Array Iterator object that contains the key/value pairs for each index in the array.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/entries

impl Array

pub fn values(&self) -> ArrayIterator

The values() method returns a new Array Iterator object that contains the values for each index in the array.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/values