Struct wasm_bindgen::js::DataView

pub struct DataView { /* fields omitted */ }

Methods

impl DataView

pub fn new(
buffer: &ArrayBuffer,
byteOffset: usize,
byteLength: usize
) -> DataView

The DataView view provides a low-level interface for reading and writing multiple number types in an ArrayBuffer irrespective of the platform's endianness.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView

impl DataView

pub fn buffer(&self) -> ArrayBuffer

The ArrayBuffer referenced by this view. Fixed at construction time and thus read only.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/buffer

impl DataView

pub fn byte_length(&self) -> usize

The length (in bytes) of this view from the start of its ArrayBuffer. Fixed at construction time and thus read only.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/byteLength

impl DataView

pub fn byte_offset(&self) -> usize

The offset (in bytes) of this view from the start of its ArrayBuffer. Fixed at construction time and thus read only.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/byteOffset

impl DataView

pub fn get_int8(&self, byte_offset: usize) -> i8

The getInt8() method gets a signed 8-bit integer (byte) at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt8

impl DataView

pub fn get_uint8(&self, byte_offset: usize) -> u8

The getUint8() method gets a unsigned 8-bit integer (byte) at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint8

impl DataView

pub fn get_int16(&self, byte_offset: usize) -> i16

The getInt16() method gets a signed 16-bit integer (byte) at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt16

impl DataView

pub fn get_uint16(&self, byte_offset: usize) -> u16

The getUint16() an unsigned 16-bit integer (unsigned byte) at the specified byte offset from the start of the view.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint16

impl DataView

pub fn get_int32(&self, byte_offset: usize) -> i32

The getInt32() method gets a signed 16-bit integer (byte) at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getInt32

impl DataView

pub fn get_uint32(&self, byte_offset: usize) -> u32

The getUint32() an unsigned 16-bit integer (unsigned byte) at the specified byte offset from the start of the view.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getUint32

impl DataView

pub fn get_float32(&self, byte_offset: usize) -> f32

The getFloat32() method gets a signed 32-bit float (float) at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getFloat32

impl DataView

pub fn get_float64(&self, byte_offset: usize) -> f64

The getFloat64() method gets a signed 32-bit float (float) at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/getFloat64

impl DataView

pub fn set_int8(&self, byte_offset: usize, value: i8)

The setInt8() method stores a signed 8-bit integer (byte) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt8

impl DataView

pub fn set_uint8(&self, byte_offset: usize, value: u8)

The setUint8() method stores an unsigned 8-bit integer (byte) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint8

impl DataView

pub fn set_int16(&self, byte_offset: usize, value: i16)

The setInt16() method stores a signed 16-bit integer (byte) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt16

impl DataView

pub fn set_uint16(&self, byte_offset: usize, value: u16)

The setUint16() method stores an unsigned 16-bit integer (byte) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint16

impl DataView

pub fn set_int32(&self, byte_offset: usize, value: i32)

The setInt32() method stores a signed 32-bit integer (byte) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setInt32

impl DataView

pub fn set_uint32(&self, byte_offset: usize, value: u32)

The setUint32() method stores an unsigned 32-bit integer (byte) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setUint32

impl DataView

pub fn set_float32(&self, byte_offset: usize, value: f32)

The setFloat32() method stores a signed 32-bit float (float) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setFloat32

impl DataView

pub fn set_float64(&self, byte_offset: usize, value: f64)

The setFloat64() method stores a signed 64-bit float (float) value at the specified byte offset from the start of the DataView.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DataView/setFloat64