Struct wasm_bindgen::js::Date

pub struct Date { /* fields omitted */ }

Methods

impl Date

pub fn get_date(&self) -> u32

The getDate() method returns the day of the month for the specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getDate

impl Date

pub fn get_day(&self) -> u32

The getDay() method returns the day of the week for the specified date according to local time, where 0 represents Sunday. For the day of the month see getDate().

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getDay

impl Date

pub fn get_full_year(&self) -> u32

The getFullYear() method returns the year of the specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getFullYear

impl Date

pub fn get_hours(&self) -> u32

The getHours() method returns the hour for the specified date, according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getHours

impl Date

pub fn get_milliseconds(&self) -> u32

The getMilliseconds() method returns the milliseconds in the specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getMilliseconds

impl Date

pub fn get_minutes(&self) -> u32

The getMinutes() method returns the minutes in the specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getMinutes

impl Date

pub fn get_month(&self) -> u32

The getMonth() method returns the month in the specified date according to local time, as a zero-based value (where zero indicates the first month of the year).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getMonth

impl Date

pub fn get_seconds(&self) -> u32

The getSeconds() method returns the seconds in the specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getSeconds

impl Date

pub fn get_time(&self) -> f64

The getTime() method returns the numeric value corresponding to the time for the specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getTime

impl Date

pub fn get_timezone_offset(&self) -> f64

The getTimezoneOffset() method returns the time zone difference, in minutes, from current locale (host system settings) to UTC.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getTimezoneOffset

impl Date

pub fn get_utc_date(&self) -> u32

The getUTCDate() method returns the day (date) of the month in the specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCDate

impl Date

pub fn get_utc_day(&self) -> u32

The getUTCDay() method returns the day of the week in the specified date according to universal time, where 0 represents Sunday.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCDay

impl Date

pub fn get_utc_full_year(&self) -> u32

The getUTCFullYear() method returns the year in the specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCFullYear

impl Date

pub fn get_utc_hours(&self) -> u32

The getUTCHours() method returns the hours in the specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCHours

impl Date

pub fn get_utc_milliseconds(&self) -> u32

The getUTCMilliseconds() method returns the milliseconds in the specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCMilliseconds

impl Date

pub fn get_utc_minutes(&self) -> u32

The getUTCMinutes() method returns the minutes in the specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCMinutes

impl Date

pub fn get_utc_month(&self) -> u32

The getUTCMonth() returns the month of the specified date according to universal time, as a zero-based value (where zero indicates the first month of the year).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCMonth

impl Date

pub fn get_utc_seconds(&self) -> u32

The getUTCSeconds() method returns the seconds in the specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getUTCSeconds

impl Date

pub fn new() -> Date

Creates a JavaScript Date instance that represents a single moment in time. Date objects are based on a time value that is the number of milliseconds since 1 January 1970 UTC.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date

impl Date

pub fn now() -> f64

The Date.now() method returns the number of milliseconds elapsed since January 1, 1970 00:00:00 UTC.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/now

impl Date

pub fn parse(date: JsString) -> f64

The Date.parse() method parses a string representation of a date, and returns the number of milliseconds since January 1, 1970, 00:00:00 UTC or NaN if the string is unrecognized or, in some cases, contains illegal date values (e.g. 2015-02-31).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/parse

impl Date

pub fn set_date(&self, day: u32) -> f64

The setDate() method sets the day of the Date object relative to the beginning of the currently set month.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setDate

impl Date

pub fn set_full_year(&self, year: u32) -> f64

The setFullYear() method sets the full year for a specified date according to local time. Returns new timestamp.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setFullYear

impl Date

pub fn set_hours(&self, hours: u32) -> f64

The setHours() method sets the hours for a specified date according to local time, and returns the number of milliseconds since January 1, 1970 00:00:00 UTC until the time represented by the updated Date instance.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setHours

impl Date

pub fn set_milliseconds(&self, milliseconds: u32) -> f64

The setMilliseconds() method sets the milliseconds for a specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setMilliseconds

impl Date

pub fn set_minutes(&self, minutes: u32) -> f64

The setMinutes() method sets the minutes for a specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setMinutes

impl Date

pub fn set_month(&self, month: u32) -> f64

The setMonth() method sets the month for a specified date according to the currently set year.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setMonth

impl Date

pub fn set_seconds(&self, seconds: u32) -> f64

The setSeconds() method sets the seconds for a specified date according to local time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setSeconds

impl Date

pub fn set_time(&self, time: f64) -> f64

The setTime() method sets the Date object to the time represented by a number of milliseconds since January 1, 1970, 00:00:00 UTC.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setTime

impl Date

pub fn set_utc_date(&self, day: u32) -> f64

The setUTCDate() method sets the day of the month for a specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCDate

impl Date

pub fn set_utc_full_year(&self, year: u32) -> f64

The setUTCFullYear() method sets the full year for a specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCFullYear

impl Date

pub fn set_utc_hours(&self, hours: u32) -> f64

The setUTCHours() method sets the hour for a specified date according to universal time, and returns the number of milliseconds since January 1, 1970 00:00:00 UTC until the time represented by the updated Date instance.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCHours

impl Date

pub fn set_utc_milliseconds(&self, milliseconds: u32) -> f64

The setUTCMilliseconds() method sets the milliseconds for a specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCMilliseconds

impl Date

pub fn set_utc_minutes(&self, minutes: u32) -> f64

The setUTCMinutes() method sets the minutes for a specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCMinutes

impl Date

pub fn set_utc_month(&self, month: u32) -> f64

The setUTCMonth() method sets the month for a specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCMonth

impl Date

pub fn set_utc_seconds(&self, seconds: u32) -> f64

The setUTCSeconds() method sets the seconds for a specified date according to universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCSeconds

impl Date

pub fn to_date_string(&self) -> JsString

The toDateString() method returns the date portion of a Date object in human readable form in American English.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toDateString

impl Date

pub fn to_iso_string(&self) -> JsString

The toISOString() method returns a string in simplified extended ISO format (ISO 8601), which is always 24 or 27 characters long (YYYY-MM-DDTHH:mm:ss.sssZ or ±YYYYYY-MM-DDTHH:mm:ss.sssZ, respectively). The timezone is always zero UTC offset, as denoted by the suffix "Z"

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toISOString

impl Date

pub fn to_json(&self) -> JsString

The toJSON() method returns a string representation of the Date object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toJSON

impl Date

pub fn to_locale_date_string(
&self,
locale: JsString,
options: JsValue
) -> JsString

The toLocaleDateString() method returns a string with a language sensitive representation of the date portion of this date. The new locales and options arguments let applications specify the language whose formatting conventions should be used and allow to customize the behavior of the function. In older implementations, which ignore the locales and options arguments, the locale used and the form of the string returned are entirely implementation dependent.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toLocaleDateString

impl Date

pub fn to_locale_string(&self, locale: JsString, options: JsValue) -> JsString

The toLocaleString() method returns a string with a language sensitive representation of this date. The new locales and options arguments let applications specify the language whose formatting conventions should be used and customize the behavior of the function. In older implementations, which ignore the locales and options arguments, the locale used and the form of the string returned are entirely implementation dependent.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toLocaleString

impl Date

pub fn to_locale_time_string(&self, locale: JsString) -> JsString

The toLocaleTimeString() method returns a string with a language sensitive representation of the time portion of this date. The new locales and options arguments let applications specify the language whose formatting conventions should be used and customize the behavior of the function. In older implementations, which ignore the locales and options arguments, the locale used and the form of the string returned are entirely implementation dependent.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toLocaleTimeString

impl Date

pub fn to_string(&self) -> JsString

The toString() method returns a string representing the specified Date object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toString

impl Date

pub fn to_time_string(&self) -> JsString

The toTimeString() method returns the time portion of a Date object in human readable form in American English.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toTimeString

impl Date

pub fn to_utc_string(&self) -> JsString

The toUTCString() method converts a date to a string, using the UTC time zone.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toUTCString

impl Date

pub fn utc(year: f64, month: f64) -> f64

The Date.UTC() method accepts the same parameters as the longest form of the constructor, and returns the number of milliseconds in a Date object since January 1, 1970, 00:00:00, universal time.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/UTC

impl Date

pub fn value_of(&self) -> Date

The valueOf() method returns the primitive value of a Date object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/valueOf