Struct wasm_bindgen::js::JsString

pub struct JsString { /* fields omitted */ }

Methods

impl JsString

pub fn length(&self) -> u32

The length property of a String object indicates the length of a string, in UTF-16 code units.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/length

impl JsString

pub fn char_at(&self, index: u32) -> JsString

The String object's charAt() method returns a new string consisting of the single UTF-16 code unit located at the specified offset into the string.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/charAt

impl JsString

pub fn char_code_at(&self, index: u32) -> f64

The charCodeAt() method returns an integer between 0 and 65535 representing the UTF-16 code unit at the given index (the UTF-16 code unit matches the Unicode code point for code points representable in a single UTF-16 code unit, but might also be the first code unit of a surrogate pair for code points not representable in a single UTF-16 code unit, e.g. Unicode code points > 0x10000). If you want the entire code point value, use codePointAt().

Returns NaN if index is out of range.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/charCodeAt

impl JsString

pub fn code_point_at(&self, pos: u32) -> JsValue

The codePointAt() method returns a non-negative integer that is the Unicode code point value.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/codePointAt

impl JsString

pub fn concat(&self, string_2: &JsString) -> JsString

The concat() method concatenates the string arguments to the calling string and returns a new string.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/concat

impl JsString

pub fn includes(&self, search_string: &JsString, position: i32) -> bool

The includes() method determines whether one string may be found within another string, returning true or false as appropriate.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/includes

impl JsString

pub fn index_of(&self, search_value: &JsString, from_index: i32) -> i32

The indexOf() method returns the index within the calling String object of the first occurrence of the specified value, starting the search at fromIndex. Returns -1 if the value is not found.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/indexOf

impl JsString

pub fn slice(&self, start: u32, end: u32) -> JsString

The slice() method extracts a section of a string and returns it as a new string, without modifying the original string.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/slice

impl JsString

pub fn starts_with(&self, search_string: &JsString, position: u32) -> bool

The startsWith() method determines whether a string begins with the characters of a specified string, returning true or false as appropriate.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/startsWith

impl JsString

pub fn substring(&self, index_start: u32, index_end: u32) -> JsString

The substring() method returns the part of the string between the start and end indexes, or to the end of the string.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/substring

impl JsString

pub fn substr(&self, start: i32, length: i32) -> JsString

The substr() method returns the part of a string between the start index and a number of characters after it.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/substr

impl JsString

pub fn to_lower_case(&self) -> JsString

The toLowerCase() method returns the calling string value converted to lower case.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/toLowerCase

impl JsString

pub fn to_string(&self) -> JsString

The toString() method returns a string representing the specified object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/toString

impl JsString

pub fn to_upper_case(&self) -> JsString

The toUpperCase() method returns the calling string value converted to uppercase (the value will be converted to a string if it isn't one).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/toUpperCase

impl JsString

pub fn trim(&self) -> JsString

The trim() method removes whitespace from both ends of a string. Whitespace in this context is all the whitespace characters (space, tab, no-break space, etc.) and all the line terminator characters (LF, CR, etc.).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/trim

impl JsString

pub fn trim_end(&self) -> JsString

The trimEnd() method removes whitespace from the end of a string. trimRight() is an alias of this method.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/trimEnd

impl JsString

pub fn trim_right(&self) -> JsString

The trimEnd() method removes whitespace from the end of a string. trimRight() is an alias of this method.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/trimEnd

impl JsString

pub fn trim_start(&self) -> JsString

The trimStart() method removes whitespace from the beginning of a string. trimLeft() is an alias of this method.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/trimStart

impl JsString

pub fn trim_left(&self) -> JsString

The trimStart() method removes whitespace from the beginning of a string. trimLeft() is an alias of this method.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/trimStart

impl JsString

pub fn value_of(&self) -> JsString

The valueOf() method returns the primitive value of a String object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/valueOf