Struct wasm_bindgen::js::Number

pub struct Number { /* fields omitted */ }

Methods

impl Number

pub fn is_finite(value: &JsValue) -> bool

The Number.isFinite() method determines whether the passed value is a finite number.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isFinite

impl Number

pub fn is_integer(value: &JsValue) -> bool

The Number.isInteger() method determines whether the passed value is an integer.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isInteger

impl Number

pub fn is_safe_integer(value: &JsValue) -> bool

The Number.isSafeInteger() method determines whether the provided value is a number that is a safe integer.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isSafeInteger

impl Number

pub fn new(value: JsValue) -> Number

The Number JavaScript object is a wrapper object allowing you to work with numerical values. A Number object is created using the Number() constructor.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number

impl Number

pub fn parse_int(text: &str, radix: u8) -> Number

The Number.parseInt() method parses a string argument and returns an integer of the specified radix or base.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/parseInt

impl Number

pub fn parse_float(text: &str) -> Number

The Number.parseFloat() method parses a string argument and returns a floating point number.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/parseFloat

impl Number

pub fn to_locale_string(&self, locale: &str) -> JsString

The toLocaleString() method returns a string with a language sensitive representation of this number.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString

impl Number

pub fn to_precision(&self, precision: u8) -> Result<JsString, JsValue>

The toPrecision() method returns a string representing the Number object to the specified precision.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toPrecision

impl Number

pub fn to_fixed(&self, digits: u8) -> Result<JsString, JsValue>

The toFixed() method returns a string representing the Number object using fixed-point notation.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toFixed

impl Number

pub fn to_exponential(&self, fraction_digits: u8) -> Result<JsString, JsValue>

The toExponential() method returns a string representing the Number object in exponential notation.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toExponential

impl Number

pub fn to_string(&self, radix: u8) -> Result<JsString, JsValue>

The toString() method returns a string representing the specified Number object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toString

impl Number

pub fn value_of(&self) -> f64

The valueOf() method returns the wrapped primitive value of a Number object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/valueOf