Struct wasm_bindgen::js::Object

pub struct Object { /* fields omitted */ }

Methods

impl Object

pub fn has_own_property(&self, property: &JsValue) -> bool

The hasOwnProperty() method returns a boolean indicating whether the object has the specified property as its own property (as opposed to inheriting it).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/hasOwnProperty

impl Object

pub fn is_extensible(object: &Object) -> bool

The Object.isExtensible() method determines if an object is extensible (whether it can have new properties added to it).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/isExtensible

impl Object

pub fn is_frozen(object: &Object) -> bool

The Object.isFrozen() determines if an object is frozen.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/isFrozen

impl Object

pub fn is_sealed(object: &Object) -> bool

The Object.isSealed() method determines if an object is sealed.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/isSealed

impl Object

pub fn is_prototype_of(&self, value: &JsValue) -> bool

The isPrototypeOf() method checks if an object exists in another object's prototype chain.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/isPrototypeOf

impl Object

pub fn keys(object: &Object) -> Array

The Object.keys() method returns an array of a given object's property names, in the same order as we get with a normal loop.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/keys

impl Object

pub fn new() -> Object

The Object constructor creates an object wrapper.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object

impl Object

pub fn prevent_extensions(object: &Object)

The Object.preventExtensions() method prevents new properties from ever being added to an object (i.e. prevents future extensions to the object).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/preventExtensions

impl Object

pub fn property_is_enumerable(&self, property: &JsValue) -> bool

The propertyIsEnumerable() method returns a Boolean indicating whether the specified property is enumerable.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/propertyIsEnumerable

impl Object

pub fn seal(value: &JsValue) -> JsValue

The Object.seal() method seals an object, preventing new properties from being added to it and marking all existing properties as non-configurable. Values of present properties can still be changed as long as they are writable.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/seal

impl Object

pub fn set_prototype_of(object: &Object, prototype: &Object) -> Object

The Object.setPrototypeOf() method sets the prototype (i.e., the internal [[Prototype]] property) of a specified object to another object or null.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/setPrototypeOf

impl Object

pub fn to_locale_string(&self) -> JsString

The toLocaleString() method returns a string representing the object. This method is meant to be overridden by derived objects for locale-specific purposes.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/toLocaleString

impl Object

pub fn to_string(&self) -> JsString

The toString() method returns a string representing the object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/toString

impl Object

pub fn value_of(&self) -> Object

The valueOf() method returns the primitive value of the specified object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/valueOf

impl Object

pub fn values(object: &Object) -> Array

The Object.values() method returns an array of a given object's own enumerable property values, in the same order as that provided by a for...in loop (the difference being that a for-in loop enumerates properties in the prototype chain as well).

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/values