Struct wasm_bindgen::js::Reflect

pub struct Reflect { /* fields omitted */ }

Methods

impl Reflect

pub fn apply(
target: &Function,
this_argument: &JsValue,
arguments_list: &Array
) -> JsValue

The static Reflect.apply() method calls a target function with arguments as specified.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/apply

impl Reflect

pub fn construct(target: &Function, arguments_list: &Array) -> JsValue

The static Reflect.construct() method acts like the new operator, but as a function. It is equivalent to calling new target(...args). It gives also the added option to specify a different prototype.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/construct

impl Reflect

pub fn construct_with_new_target(
target: &Function,
arguments_list: &Array,
new_target: &Function
) -> JsValue

impl Reflect

pub fn define_property(
target: &Object,
property_key: &JsValue,
attributes: &Object
) -> bool

The static Reflect.defineProperty() method is like Object.defineProperty() but returns a Boolean.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/defineProperty

impl Reflect

pub fn delete_property(target: &Object, key: &JsValue) -> bool

The static Reflect.deleteProperty() method allows to delete properties. It is like the delete operator as a function.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/deleteProperty

impl Reflect

pub fn get(target: &Object, key: &JsValue) -> JsValue

The static Reflect.get() method works like getting a property from an object (target[propertyKey]) as a function.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/get

impl Reflect

pub fn get_own_property_descriptor(
target: &Object,
property_key: &JsValue
) -> JsValue

The static Reflect.getOwnPropertyDescriptor() method is similar to Object.getOwnPropertyDescriptor(). It returns a property descriptor of the given property if it exists on the object, undefined otherwise.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/getOwnPropertyDescriptor

impl Reflect

pub fn get_prototype_of(target: &Object) -> Object

The static Reflect.getPrototypeOf() method is almost the same method as Object.getPrototypeOf(). It returns the prototype (i.e. the value of the internal [[Prototype]] property) of the specified object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/getPrototypeOf

impl Reflect

pub fn has(target: &Object, property_key: &JsValue) -> bool

The static Reflect.has() method works like the in operator as a function.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/has

impl Reflect

pub fn is_extensible(target: &Object) -> bool

The static Reflect.isExtensible() method determines if an object is extensible (whether it can have new properties added to it). It is similar to Object.isExtensible(), but with some differences.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/isExtensible

impl Reflect

pub fn own_keys(target: &Object) -> Array

The static Reflect.ownKeys() method returns an array of the target object's own property keys.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/ownKeys

impl Reflect

pub fn prevent_extensions(target: &Object) -> bool

The static Reflect.preventExtensions() method prevents new properties from ever being added to an object (i.e. prevents future extensions to the object). It is similar to Object.preventExtensions(), but with some differences.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/preventExtensions

impl Reflect

pub fn set(target: &Object, property_key: &JsValue, value: &JsValue) -> bool

The static Reflect.set() method works like setting a property on an object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/set

impl Reflect

pub fn set_with_receiver(
target: &Object,
property_key: &JsValue,
value: &JsValue,
receiver: &Object
) -> bool

impl Reflect

pub fn set_prototype_of(target: &Object, prototype: &JsValue) -> bool

The static Reflect.setPrototypeOf() method is the same method as Object.setPrototypeOf(). It sets the prototype (i.e., the internal [[Prototype]] property) of a specified object to another object or to null.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Reflect/setPrototypeOf