Struct wasm_bindgen::js::Set

pub struct Set { /* fields omitted */ }

Methods

impl Set

pub fn add(&self, value: &JsValue) -> Set

The add() method appends a new element with a specified value to the end of a Set object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set/add

impl Set

pub fn clear(&self)

The clear() method removes all elements from a Set object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set/clear

impl Set

pub fn delete(&self, value: &JsValue) -> bool

The delete() method removes the specified element from a Set object.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set/delete

impl Set

pub fn has(&self, value: &JsValue) -> bool

The has() method returns a boolean indicating whether an element with the specified value exists in a Set object or not.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set/has

impl Set

pub fn new() -> Set

The Set object lets you store unique values of any type, whether primitive values or object references.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set

impl Set

pub fn size(&self) -> u32

The size accessor property returns the number of elements in a Set object.

https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Set/size

impl Set

pub fn entries(&self) -> SetIterator

The entries() method returns a new Iterator object that contains an array of [value, value] for each element in the Set object, in insertion order. For Set objects there is no key like in Map objects. However, to keep the API similar to the Map object, each entry has the same value for its key and value here, so that an array [value, value] is returned.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set/entries

impl Set

pub fn keys(&self) -> SetIterator

The keys() method is an alias for this method (for similarity with Map objects); it behaves exactly the same and returns values of Set elements.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set/values

impl Set

pub fn values(&self) -> SetIterator

The values() method returns a new Iterator object that contains the values for each element in the Set object in insertion order.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set/values