Struct wasm_bindgen::js::Uint8Array

pub struct Uint8Array { /* fields omitted */ }

Methods

impl Uint8Array

pub fn new(constructor_arg: JsValue) -> Uint8Array

The Uint8Array() constructor creates an array of unsigned 8-bit integers.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Uint8Array

impl Uint8Array

pub fn fill(&self, value: JsValue, start: u32, end: u32) -> Uint8Array

The fill() method fills all the elements of an array from a start index to an end index with a static value. The end index is not included.

http://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray/fill