wasm_bindgen::sys

Struct Null

pub struct Null { /* private fields */ }

The JavaScript null value.

This type represents the JavaScript null primitive value and can be used as a generic type parameter to indicate that a value is null.

Implementations

impl Null

pub const NULL: Null

The null constant.