wasm_bindgen

Trait ErasableGeneric

pub unsafe trait ErasableGeneric {
    type Repr: 'static;
}

Private marker trait for erasable generics - types with this trait have the same repr for all generic param values, and can therefore be transmuted on the singular Repr type representation on ABI boundaries.

Safety

This type must only be implemented on types known to be repr equivalent to their Repr type.

Required Associated Types

type Repr: 'static

The singular concrete type that all generic variants can be transmuted on