Crate wasm_bindgen

Runtime support for the wasm-bindgen tool

This crate contains the runtime support necessary for wasm-bindgen the attribute and tool. Crates pull in the #[wasm_bindgen] attribute through this crate and this crate also provides JS bindings through the JsValue interface.

Modules

closure
Support for long-lived closures in wasm-bindgen

convert
This is mostly an internal module, no stability guarantees are provided. Use at your own risk.

js
Bindings to JavaScript's standard, built-in objects, including their methods and properties.

prelude
A module which is typically glob imported from:

Structs

JsStatic
Wrapper type for imported statics.

JsValue
Representation of an object owned by JS.

Functions

throw
Throws a JS exception.