Struct wasm_bindgen::closure::Closure

pub struct Closure<T: ?Sized> { /* fields omitted */ }

A handle to both a closure in Rust as well as JS closure which will invoke the Rust closure.

A Closure is the primary way that a 'static lifetime closure is transferred from Rust to JS. Closure currently requires that the closures it's created with have the 'static lifetime in Rust for soundness reasons.

This type is a "handle" in the sense that whenever it is dropped it will invalidate the JS closure that it refers to. Any usage of the closure in JS after the Closure has been dropped will raise an exception. It's then up to you to arrange for Closure to be properly deallocate at an appropriate location in your program.

The type parameter on Closure is the type of closure that this represents. Currently this can only be the Fn and FnMut traits with up to 7 arguments (and an optional return value). The arguments/return value of the trait must be numbers like u32 for now, although this restriction may be lifted in the future!

Example

#[wasm_bindgen]
extern {
    fn setTimeout(closure: &Closure<FnMut()>, time: u32);

    #[wasm_bindgen(js_namespace = console)]
    fn log(s: &str);
}

#[wasm_bindgen]
pub struct ClosureHandle(Closure<FnMut()>);

#[wasm_bindgen]
pub fn run() -> ClosureHandle {
    // First up we use `Closure::new` to wrap up a Rust closure and create
    a JS closure.
    let cb = Closure::new(|| {
        log("timeout elapsed!");
    });

    // Next we pass this via reference to the `setTimeout` function, and
    // `setTimeout` gets a handle to the corresponding JS closure.
    setTimeout(&cb, 1_000);

    // If we were to drop `cb` here it would cause an exception to be raised
    // when the timeout elapses. Instead we *return* our handle back to JS
    // so JS can tell us later when it would like to deallocate this handle.
    ClosureHandle(cb)
}

Methods

impl<T> Closure<T> where
T: ?Sized,

pub fn new<F>(t: F) -> Closure<T> where
F: Unsize<T> + 'static,

Creates a new instance of Closure from the provided Rust closure.

Note that the closure provided here, F, has a few requirements associated with it:

It must implement Fn or FnMut

It must be 'static, aka no stack references (use the move keyword)

It can have at most 7 arguments

Its arguments and return values are all wasm types like u32/f64.

This is unfortunately pretty restrictive for now but hopefully some of these restrictions can be lifted in the future!

pub fn wrap(t: Box<T>) -> Closure<T>

A mostly internal function to wrap a boxed closure inside a Closure type.

This is the function where the JS closure is manufactured.

pub fn forget(self)

Leaks this Closure to ensure it remains valid for the duration of the entire program.

Note: this function will leak memory. It should be used sparingly to ensure the memory leak doesn't affect the program too much.

When a Closure is dropped it will invalidate the associated JS closure, but this isn't always desired. Some callbacks are alive for the entire duration of the program, so this can be used to conveniently leak this instance of Closure while performing as much internal cleanup as it can.