Trait wasm_bindgen::convert::OptionFromWasmAbi

pub trait OptionFromWasmAbi: FromWasmAbi {
    fn is_none(abi: &Self::Abi) -> bool;
}

Indicates that this type can be received from JS as Option<Self>.

This trait is used when implementing FromWasmAbi for Option<T>.

Required Methods

fn is_none(abi: &Self::Abi) -> bool

Tests whether the argument is a “none” instance. If so it will be deserialized as None, and otherwise it will be passed to FromWasmAbi.