Trait wasm_bindgen::convert::OptionIntoWasmAbi

pub trait OptionIntoWasmAbi: IntoWasmAbi {
    fn none() -> Self::Abi;
}

Indicates that this type can be passed to JS as Option<Self>.

This trait is used when implementing IntoWasmAbi for Option<T>.

Required methods

fn none() -> Self::Abi

Returns an ABI instance indicating "none", which JS will interpret as the None branch of this option.

It should be guaranteed that the IntoWasmAbi can never produce the ABI value returned here.

Loading content...