Trait wasm_bindgen::convert::RefFromWasmAbi

pub trait RefFromWasmAbi: WasmDescribe {
    type Abi: WasmAbi;
    type Anchor: Deref<Target = Self>;

    unsafe fn ref_from_abi(js: Self::Abi) -> Self::Anchor;
}

A trait for anything that can be recovered as some sort of shared reference from the wasm ABI boundary.

This is the shared reference variant of the opposite operation as IntoWasmAbi.

Required Associated Types

type Abi: WasmAbi

The wasm ABI type references to Self are recovered from.

type Anchor: Deref<Target = Self>

The type that holds the reference to Self for the duration of the invocation of the function that has an &Self parameter. This is required to ensure that the lifetimes don’t persist beyond one function call, and so that they remain anonymous.

Required Methods

unsafe fn ref_from_abi(js: Self::Abi) -> Self::Anchor

Recover a Self::Anchor from Self::Abi.

Safety

Same as FromWasmAbi::from_abi.