Trait wasm_bindgen::convert::WasmAbi

pub unsafe trait WasmAbi { }

An unsafe trait which represents types that are ABI-safe to pass via wasm arguments.

This is an unsafe trait to implement as there’s no guarantee the type is actually safe to transfer across the was boundary, it’s up to you to guarantee this so codegen works correctly.