Function wasm_bindgen::throw_str

pub fn throw_str(s: &str) -> !

Throws a JS exception.

This function will throw a JS exception with the message provided. The function will not return as the wasm stack will be popped when the exception is thrown.

Note that it is very easy to leak memory with this function because this function, unlike panic! on other platforms, will not run destructors. It's recommended to return a Result where possible to avoid the worry of leaks.